requirements.txt
streamlit 
langchain 
langchain_ollama
selenium
beautifulsoup4
lxml 
html5lib
python-dotenv
requests
python-dotenv 
pandas 
numpy 
scikit-learn 
xgboost